승인 버튼 클릭시 그 기업에게 쿠폰 생성

+++ src/main/java/com/example/bookangel/beans/dao/CouponDAO.java
@@ -17,7 +17,7 @@ public class CouponDAO {
     private final CouponMapper couponMapper;
 
     // 쿠폰 사용 가능여부 확인
-    public Integer checkCoupon(CouponVO couponVO){
+    public Long checkCoupon(CouponVO couponVO){
         log.info("[DAO]-쿠폰 확인...............");
         return couponMapper.checkCoupon(couponVO);
     }
@@ -25,23 +25,23 @@ public class CouponDAO {
     // 쿠폰 사용
     public boolean useCoupon(CouponVO couponVO){
         log.info("[DAO]-쿠폰 사용...............");
-        return couponMapper.useCoupon(couponVO) == 1;
+        return couponMapper.useCoupon(couponVO) == 1L;
     }
 
     // 기업 쿠폰 리스트 조회
-    public List<CouponVO> companyCouponList(int memberNum){
+    public List<CouponVO> companyCouponList(long memberNum){
         log.info("[DAO]-기업 쿠폰 리스트 확인...............");
         return couponMapper.companyCouponList(memberNum);
     }
 
     public boolean isExist(String couponName){
         log.info("[DAO]-쿠폰이 존재하는지...............");
-        return couponMapper.isExist(couponName) == 1;
+        return couponMapper.isExist(couponName) == 1L;
     }
 
     // 쿠폰만들기
     public boolean makeCoupon(CouponVO couponVO){
         log.info("[DAO]-쿠폰 생성...............");
-        return couponMapper.makeCoupon(couponVO) == 1;
+        return couponMapper.makeCoupon(couponVO) == 1L;
     }
 }
