+++ src/main/java/com/example/bookangel/mappers/CouponMapper.java
@@ -8,18 +8,18 @@ import java.util.List;
 @Mapper
 public interface CouponMapper {
     // 사용 가능한 쿠폰인지
-    public Integer checkCoupon(CouponVO couponVO);
+    public Long checkCoupon(CouponVO couponVO);
 
     // 쿠폰 사용
-    public int useCoupon(CouponVO couponVO);
+    public long useCoupon(CouponVO couponVO);
 
     // 기업 쿠폰 리스트 조회
-    public List<CouponVO> companyCouponList(int memberNum);
+    public List<CouponVO> companyCouponList(long memberNum);
 
     // 쿠폰이 있는지
-    public int isExist(String couponName);
+    public long isExist(String couponName);
 
     // 쿠폰만들기
-    public int makeCoupon(CouponVO couponVO);
+    public long makeCoupon(CouponVO couponVO);
 }
